+++ src/Pagination/PaginationArgs.php
@@ -45,7 +45,7 @@ class PaginationArgs
 
         if ($instance->first <= 0) {
             throw new Error(
-                "Requested pagination amount must be more than 0, got $instance->first"
+                self::requestedZeroOrLessItems($instance->first)
             );
         }
 
@@ -55,11 +55,21 @@ class PaginationArgs
             && $instance->first > $paginateMaxCount
         ) {
             throw new Error(
-                "Maximum number of {$paginateMaxCount} requested items exceeded. Fetch smaller chunks."
+                self::requestedTooManyItems($paginateMaxCount, $instance->first)
             );
         }
 
         return $instance;
+    }
+
+    public static function requestedZeroOrLessItems(int $amount): string
+    {
+        return "Requested pagination amount must be more than 0, got {$amount}.";
+    }
+
+    public static function requestedTooManyItems(int $maxCount, int $actualCount): string
+    {
+        return "Maximum number of {$maxCount} requested items exceeded, got {$actualCount}. Fetch smaller chunks.";
     }
 
     /**